requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
ffmpeg
pandas
numpy
email-validator
openai-whisper

librosa
soundfile

groq